+++ packages/grafana-ui/src/components/Cascader/Cascader.tsx
@@ -4,7 +4,7 @@ import RCCascader from 'rc-cascader';
 
 import { Select } from '../Select/Select';
 import { FormInputSize } from '../Forms/types';
-import { Input } from '../Forms/Input/Input';
+import { Input } from '../Input/Input';
 import { SelectableValue } from '@grafana/data';
 import { css } from 'emotion';
 import { onChangeCascader } from './optionMappings';
