+++ packages/grafana-ui/src/components/ColorPicker/ColorInput.tsx
@@ -3,7 +3,7 @@ import tinycolor from 'tinycolor2';
 import debounce from 'lodash/debounce';
 
 import { ColorPickerProps } from './ColorPickerPopover';
-import { Input } from '../Input/Input';
+import { Input } from '../Forms/Legacy/Input/Input';
 
 interface ColorInputState {
   previousColor: string;
