+++ packages/grafana-ui/src/components/Forms/Field.story.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useCallback } from 'react';
 import { boolean, number, text } from '@storybook/addon-knobs';
 import { Field } from './Field';
-import { Input } from './Input/Input';
+import { Input } from '../Input/Input';
 import { Switch } from './Switch';
 import mdx from './Field.mdx';
