+++ packages/grafana-ui/src/components/Forms/Form.story.tsx
@@ -4,7 +4,7 @@ import { Legend } from './Legend';
 import { withCenteredStory } from '../../utils/storybook/withCenteredStory';
 import { withStoryContainer } from '../../utils/storybook/withStoryContainer';
 import { Field } from './Field';
-import { Input } from './Input/Input';
+import { Input } from '../Input/Input';
 import { Button } from '../Button';
 import { Form } from './Form';
 import { Switch } from './Switch';
