+++ packages/grafana-ui/src/components/Forms/getFormStyles.ts
@@ -5,7 +5,7 @@ import { getLegendStyles } from './Legend';
 import { getFieldValidationMessageStyles } from './FieldValidationMessage';
 import { getButtonStyles, ButtonVariant } from '../Button';
 import { ComponentSize } from '../../types/size';
-import { getInputStyles } from './Input/Input';
+import { getInputStyles } from '../Input/Input';
 import { getSwitchStyles } from './Switch';
 import { getCheckboxStyles } from './Checkbox';
