+++ packages/grafana-ui/src/components/OptionsUI/number.tsx
@@ -5,7 +5,7 @@ import {
   toFloatOrUndefined,
   NumberFieldConfigSettings,
 } from '@grafana/data';
-import { Input } from '../Forms/Input/Input';
+import { Input } from '../Input/Input';
 
 export const NumberValueEditor: React.FC<FieldConfigEditorProps<number, NumberFieldConfigSettings>> = ({
   value,
