+++ packages/grafana-ui/src/components/Select/IndicatorsContainer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useTheme } from '../../themes/ThemeContext';
-import { getInputStyles } from '../Forms/Input/Input';
+import { getInputStyles } from '../Input/Input';
 import { cx, css } from 'emotion';
 
 export const IndicatorsContainer = React.forwardRef<HTMLDivElement, React.PropsWithChildren<any>>((props, ref) => {
