+++ packages/grafana-ui/src/components/Select/InputControl.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTheme } from '../../themes/ThemeContext';
 import { getFocusCss, sharedInputStyle } from '../Forms/commonStyles';
-import { getInputStyles } from '../Forms/Input/Input';
+import { getInputStyles } from '../Input/Input';
 import { cx, css } from 'emotion';
 import { stylesFactory } from '../../themes';
 import { GrafanaTheme } from '@grafana/data';
