+++ packages/grafana-ui/src/components/ValueMappingsEditor/LegacyMappingRow.tsx
@@ -2,7 +2,7 @@ import React, { ChangeEvent, PureComponent } from 'react';
 
 import { FormField } from '../FormField/FormField';
 import { FormLabel } from '../FormLabel/FormLabel';
-import { Input } from '../Input/Input';
+import { Input } from '../Forms/Legacy/Input/Input';
 import { Select } from '../Forms/Legacy/Select/Select';
 
 import { MappingType, ValueMapping } from '@grafana/data';
